refactoring api

+++ src/main/java/com/dmiit3iy/reminder/service/ReminderService.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface ReminderService {
 
-    void add(Reminder reminder, int userID);
+    void add(Reminder reminder, long userID);
 
     /**
      * Получение списка напоминаний по ID
@@ -18,34 +18,34 @@ public interface ReminderService {
      * @param id
      * @return
      */
-    Reminder get(long id, int userID);
+    Reminder get(long id, long userID);
 
     /**
      * Получение списка всех напоминаний
      *
      * @return
      */
-    List<Reminder> get(int userID);
+    List<Reminder> get(long userID);
 
     /**
      * Получение списка с пагинацией
      * @param pageable
      * @return
      */
-    Page<Reminder> get(Pageable pageable, int userID);
+    Page<Reminder> get(Pageable pageable, long userID);
 
     /**
      * Получение списка с пагинацией с указанием параметров
      * @param
      * @return
      */
-    Page<Reminder> get(int page, int size, int userID);
+    Page<Reminder> get(int page, int size, long userID);
     /**
      * Получение последнего созданного напоминаия
      *
      * @return
      */
-    Reminder getLast(int userID);
+    Reminder getLast(long userID);
 
     /**
      * Получение списка напоминаний по заголовку
@@ -53,7 +53,7 @@ public interface ReminderService {
      * @param title
      * @return
      */
-    List<Reminder> get(String title, int userID);
+    List<Reminder> get(String title, long userID);
 
     /**
      * Получение списка напоминаний по описанию
@@ -62,7 +62,7 @@ public interface ReminderService {
      * @return
      */
 
-    List<Reminder> getByDescription(String description, int userID);
+    List<Reminder> getByDescription(String description, long userID);
 
     /**
      * Получение списка напоминаний по дате
@@ -71,7 +71,7 @@ public interface ReminderService {
      * @return
      */
 
-    List<Reminder> get(LocalDate localDate, int userID);
+    List<Reminder> get(LocalDate localDate, long userID);
 
     /**
      * Получение списка напоминаний по времени
@@ -79,7 +79,7 @@ public interface ReminderService {
      * @param localTime
      * @return
      */
-    List<Reminder> get(LocalTime localTime, int userID);
+    List<Reminder> get(LocalTime localTime, long userID);
 
     /**
      * Удаление напоминания по ID
@@ -87,13 +87,13 @@ public interface ReminderService {
      * @param id
      * @return
      */
-    Reminder delete(long id, int userID);
+    Reminder delete(long id, long userID);
 
     /**
      * Удаление последнего напоминания
      * @return
      */
-    Reminder delete(int userID);
+    Reminder delete(long userID);
 
     /**
      * Обновление напоминания
@@ -101,6 +101,6 @@ public interface ReminderService {
      * @param reminder
      * @return
      */
-    Reminder update(Reminder reminder, int userID);
+    Reminder update(Reminder reminder, long userID);
 
 }
